fastapi/requirements.txt
accelerate>=0.27.2
annotated-types>=0.6.0
anyio>=4.3.0
certifi>=2024.2.2
charset-normalizer>=3.3.2
click>=8.1.7
diffusers>=0.26.3
fastapi>=0.110.0
filelock>=3.13.1
fsspec>=2024.2.0
h11>=0.14.0
huggingface-hub>=0.21.3
idna>=3.6
importlib-metadata>=7.0.1
Jinja2>=3.1.3
MarkupSafe>=2.1.5
mpmath>=1.3.0
networkx>=3.2.1
numpy>=1.26.4
packaging>=23.2
pillow>=10.4.0
psutil>=5.9.8
pydantic>=2.6.3
pydantic_core>=2.16.3
python-multipart>=0.0.9
PyYAML>=6.0.1
regex>=2023.12.25
requests>=2.31.0
safetensors>=0.4.2
sniffio>=1.3.1
starlette>=0.36.3
sympy>=1.12
tokenizers>=0.15.2
# PyTorch with CUDA 12.4 support
# NOTE: PyTorch must be installed with CUDA support using the command below
# Run this command BEFORE installing other dependencies:
# python -m pip install --index-url https://download.pytorch.org/whl/cu124 torch torchvision torchaudio
# Then run: pip install -r requirements.txt
tqdm>=4.66.2
transformers>=4.38.2
typing_extensions>=4.10.0
urllib3>=2.2.1
uvicorn>=0.27.1
zipp>=3.17.0